wildberries/requirements.txt
python-telegram-bot==20.3
SQLAlchemy==1.4.46
aiohttp==3.8.4
beautifulsoup4==4.12.2
playwright==1.33.0
pandas==1.5.3
openpyxl==3.1.2
python-dateutil==2.8.2
cachetools==5.3.0
APScheduler==3.10.1
python-dotenv==1.0.0